helpers.h: fix check for st_birthtimespec on macOS
This should be based on system version, not bitness of the arch.
Notice, this condition is true for all released macOS versions,
but not for some pre-release builds of 10.6, where st_birthtimespec
is missing. There is no easy way to check for those; if you build
on 10a96 or 10a190, please modify the macro accordingly.